Move Aggregator. Define basic version of polymorphic view
